Extension field feature verifier changes

+++ common/src/proof.rs
@@ -28,7 +28,7 @@ pub struct Context {
     pub lde_domain_depth: u8,
     pub ce_blowup_factor: u8,
     pub field_modulus_bytes: Vec<u8>,
-    pub field_extension_factor: u32,
+    pub field_extension_factor: u8,
     pub options: ProofOptions,
 }
 
@@ -104,7 +104,7 @@ impl StarkProof {
             - 1;
 
         // field_modulus_bits * field_extension_factor - log2(extended trace length)
-        let max_fri_security = field_modulus_bits * self.context.field_extension_factor
+        let max_fri_security = field_modulus_bits * self.context.field_extension_factor as u32
             - self.context.lde_domain_depth as u32;
 
         std::cmp::min(std::cmp::min(result, max_fri_security), cr_security)
